cloud-run-worker/requirements.txt
Flask==3.0.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
google-cloud-dlp==3.15.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
gunicorn==21.2.0